feat:  an intermediate safepoint
I am still having problems getting the markers to filter when the
venue search list filters.

I am making this commit to preserve some of the various attempts to fix
it.

The result of my filter in SearchList is what I want to use to filter
my markers........but i cannot export this array to the state in
App.js in order to use it.  I keep putting React into an infinite loop.

Intention now is to try and replicate the same filter method in App.js
with the existing markers array that is used to initiate the displaying
markers in the first place.