feat: markers are filtering but not properly
Markers do not appear on the map initially.

As you enter characters in the text box, markers do appear, but they
are always behind the onChange event by one event?

Also, on page loading the filteredVenues array holds only 30 objects,
rather than the 30 imported from Foursquare that can be seen in the
markers array.

I guess it must be something to do with the React lifecycle?