fix: markers filter at the same time as locations list
Bug previously was that whilst the locations list filtered correctly,
the associated markers on the map were always out of sync by one key
press in the input field.

This has caused many hours of trying work arounds thinking it was a
setState asynchronisity issue, based on stackoverflow answers.

In the end, it was simply that I had used this.state.query instead of
'query' in my if statement for the filter.

This caused the lag!